cgroup_fj: function2
Check if the cgroup exists. The cgroup
will be mounted (and umounted) only if it doesn't exist at
the start of the test, otherwise the mount point will be used
instead of mouting the cgroup again (which is not possible anymore).
Removed testcases 3-6 because they tested a no longer supported
feature.
Add tests for every subsystem instead of only debug except for
testcases 1 and 2.

Signed-off-by: Cedric Hnyda <chnyda@suse.com>