controllers/cgroup_fj: remove broken testcase
cgroup_fj_function_cpuset_yes_no_1_yes_2_3_2_1 seems to have never worked.
It tried to write in the file cpus but there is no such file,
cpuset.cpus should be used in all cases.
The test was removed from runtest/controllers.
The option noprefix was removed for all function testcases since it was only
used for that testcase.

Signed-off-by: Cedric Hnyda <cedric@hnyda.net>